Return the narrowed value in the result, and remove properties and items from object and array types
