+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "@jsonhero/json-infer-types",
-  "version": "0.1.0",
+  "version": "1.0.0",
   "description": "Infer the types of a JSON value",
   "homepage": "https://github.com/jsonhero-io/json-infer-types",
   "bugs": {
