+++ src/formats/tld.ts
@@ -1,4 +1,6 @@
-import { JSONTLDFormat } from "../@types";
+export type JSONTLDFormat = {
+  name: "tld";
+};
 
 export function inferTld(value: string): JSONTLDFormat | undefined {
   const tldRegex = /^\.\w{2,14}(\.\w{2,14})*$/i;
